refactored and went through pomdp_file comments
