+++ grl/mi.py
@@ -5,7 +5,7 @@ from jax.nn import softmax
 from tqdm import trange
 
 from grl.analytical_agent import AnalyticalAgent
-from grl.policy_eval import lambda_discrep_measures
+from grl.utils.lambda_discrep import lambda_discrep_measures
 from grl.mdp import AbstractMDP, MDP
 from grl.memory import memory_cross_product
 from grl.utils.math import glorot_init, greedify
