+++ grl/utils/__init__.py
@@ -1,6 +1,6 @@
 from .fs import *
 from .math import *
 from .loss import *
-from .pe import *
+from .mdp import *
 
 RTOL = 1e-3
