+++ grl/utils/loss.py
@@ -1,7 +1,9 @@
 import jax.numpy as jnp
-from jax import jit, value_and_grad, nn
-from .pe import analytical_pe
-from grl.memory import functional_memory_cross_product, memory_cross_product
+from jax import nn
+
+from grl.utils.mdp import functional_get_occupancy, get_p_s_given_o, functional_create_td_model
+from grl.utils.pe import analytical_pe, functional_solve_mdp
+from grl.memory import functional_memory_cross_product
 
 def mem_diff(value_type: str, mem_params: jnp.ndarray, gamma: float, pi: jnp.ndarray,
              T: jnp.ndarray, R: jnp.ndarray, phi: jnp.ndarray, p0: jnp.ndarray):
@@ -61,3 +63,31 @@ def pi_discrep_q_abs_loss(pi_params: jnp.ndarray, gamma: float, T: jnp.ndarray, 
     diff, mc_vals, td_vals, pi = pi_calc_diff('q', pi_params, gamma, T, R, phi, p0)
     diff = diff * pi.T
     return jnp.abs(diff).mean(), (mc_vals, td_vals)
+
+
+def mem_abs_td_loss(mem_params: jnp.ndarray, gamma: float, pi: jnp.ndarray, T: jnp.ndarray,
+                    R: jnp.ndarray, phi: jnp.ndarray, p0: jnp.ndarray):
+    """
+    Absolute TD error loss.
+    This is an upper bound on absolute lambda discrepancy.
+    """
+    T_mem = nn.softmax(mem_params, axis=-1)
+    T_x, R_x, p0_x, phi_x = functional_memory_cross_product(T, T_mem, phi, R, p0)
+
+    # observation policy, but expanded over states
+    pi_state = phi_x @ pi
+    occupancy = functional_get_occupancy(pi_state, T_x, p0_x, gamma)
+
+    p_pi_of_s_given_o = get_p_s_given_o(phi_x, occupancy)
+
+    # TD
+    T_obs_obs, R_obs_obs = functional_create_td_model(p_pi_of_s_given_o, phi_x, T_x, R_x)
+    td_v_vals, td_q_vals = functional_solve_mdp(pi, T_obs_obs, R_obs_obs, gamma)
+    td_vals = {'v': td_v_vals, 'q': td_q_vals}
+
+    # Get starting obs distribution
+    obs_p0_x = phi_x * p0_x
+    # based on our TD model, get our observation occupancy
+    obs_occupancy = functional_get_occupancy(pi, T_obs_obs, obs_p0_x, gamma)
+
+    raise NotImplementedError
