+++ grl/vi.py
@@ -5,7 +5,8 @@ from typing import Tuple
 from functools import partial
 
 from grl.utils.math import glorot_init
-from grl.utils.pe import functional_get_occupancy, get_p_s_given_o, functional_solve_mdp, functional_create_td_model
+from grl.utils.mdp import functional_get_occupancy, get_p_s_given_o, functional_create_td_model
+from grl.utils import functional_solve_mdp
 
 @partial(jit, static_argnames=['gamma'])
 def value_iteration_step(vp: jnp.ndarray, T: jnp.ndarray, R: jnp.ndarray, gamma: float):
