+++ tests/test_grad.py
@@ -78,3 +78,6 @@ def test_example_7_m():
     memory_grad, _ = pe_grad(spec, pi, 'm', lr=1)
 
     assert np.allclose(memory_end, softmax(memory_grad, axis=-1), atol=1e-2)
+
+if __name__ == "__main__":
+    test_example_7_m()
